restructure webapp, add option to save trained mln, fix call for files
