+++ webapps/prac/webrob/pages/routes.py
@@ -48,4 +48,5 @@ def register_routes():
     from webrob.pages import log
     from webrob.pages import meshes
     from webrob.pages import views
+    from webrob.pages import utils
