+++ webapps/prac/webrob/templates/learn.html
@@ -42,6 +42,7 @@
 <div id='visualization'>
 <h1>Visualization</h1>
 <div id='viz'>
+<button onclick="mlnFileName()")>Save MLN to file</button>
 <textarea id="result" name="result">{{ result.res }}</textarea>
 </div>
 </div>
